Transaction functionality done
	new file:   conf/web.xml
    Servlet mappings

	modified:   src/html/group-information-box.jsp
    Page-loading, servlet-transaction functionality

	modified:   src/java/org/tinkerbeast/gems/servlet/InfoProvider.java
    Information formats changed

	modified:   src/java/org/tinkerbeast/gems/servlet/ProcessTransaction.java
    Transaction logic updated

	modified:   src/js/transaction-category.js
    Consolidated data returning mechanism

	modified:   src/js/transaction-unit.js
    Consolidated data returning mechanism

+++ src/js/transaction-unit.js
@@ -44,7 +44,7 @@ TransactionUnit.prototype.isValid = function() {
 
 TransactionUnit.prototype.getValue = function() {
     // TODO use val() method
-    return parseInt(this.element.children("input").val());
+    return parseFloat(this.element.children("input").val());
 };
 
 TransactionUnit.prototype.getHtmlElement = function() {
